Conflicts:
	CMakeLists.txt
	src/core/query.cpp
	src/file/autotest/basicindexingqueuetest.cpp
	src/file/autotest/fileindexerconfigutils.h
	src/file/autotest/fileindexingjob/extractor.cpp
	src/file/autotest/metadatamovertest.cpp
	src/file/cleaner/main.cpp
	src/file/extractor/app.cpp
	src/file/filewatch.cpp
	src/file/kcm/folderselectionwidget.cpp
	src/file/kcm/kcm.cpp
	src/file/kinotify.cpp
	src/file/lib/autotests/db.cpp
	src/file/lib/autotests/filefetchjobtest.cpp
	src/file/lib/autotests/filemodifyjobtest.cpp
	src/file/lib/db.cpp
	src/file/lib/xattrdetector.cpp
	src/file/main.cpp
	src/file/metadatamover.cpp
	src/file/regexpcache.cpp
	src/file/search/filesearchstore.cpp
	src/kioslaves/timeline/kio_timeline.cpp
	src/kioslaves/timeline/timelinetools.cpp
	src/pim/agent/CMakeLists.txt
	src/pim/agent/agent.cpp
	src/pim/agent/calendarindexer.h
	src/pim/akonadiplugin/searchplugin.cpp
	src/pim/lib/contactcompleter.cpp
	src/pim/lib/contactquery.cpp
	src/pim/lib/emailquery.cpp
	src/pim/lib/notequery.cpp
	src/pim/lib/query.cpp
	src/pim/search/pimsearchstore.cpp
	src/runner/baloosearchrunner.cpp
	src/tools/balooctl/main.cpp
	src/tools/baloosearch/main.cpp
	src/tools/balooshow/main.cpp

+++ src/pim/agent/akonotesindexer.cpp
@@ -26,15 +26,28 @@
 
 
 AkonotesIndexer::AkonotesIndexer(const QString& path)
-    : AbstractIndexer(), m_termGen( 0 )
+    : AbstractIndexer(), m_db(0), m_termGen( 0 )
 {
-    m_db = new Xapian::WritableDatabase(path.toUtf8().constData(), Xapian::DB_CREATE_OR_OPEN);
+    try {
+        m_db = new Xapian::WritableDatabase(path.toUtf8().constData(), Xapian::DB_CREATE_OR_OPEN);
+    }
+    catch (const Xapian::DatabaseCorruptError& err) {
+        kError() << "Database Corrupted - What did you do?";
+        kError() << err.get_error_string();
+        m_db = 0;
+    }
+    catch (const Xapian::Error &e) {
+        kError() << QString::fromStdString(e.get_type()) << QString::fromStdString(e.get_description());
+        m_db = 0;
+    }
 }
 
 AkonotesIndexer::~AkonotesIndexer()
 {
-    m_db->commit();
-    delete m_db;
+    if (m_db) {
+        m_db->commit();
+        delete m_db;
+    }
 }
 
 QStringList AkonotesIndexer::mimeTypes() const
@@ -44,6 +57,8 @@ QStringList AkonotesIndexer::mimeTypes() const
 
 void AkonotesIndexer::index(const Akonadi::Item &item)
 {
+    if (!m_db)
+        return;
     KMime::Message::Ptr msg;
     try {
         msg = item.payload<KMime::Message::Ptr>();
@@ -125,11 +140,14 @@ void AkonotesIndexer::processPart(KMime::Content* content, KMime::Content* mainC
 
 void AkonotesIndexer::commit()
 {
-    m_db->commit();
+    if (m_db)
+        m_db->commit();
 }
 
 void AkonotesIndexer::remove(const Akonadi::Item &item)
 {
+    if (!m_db)
+        return;
     try {
         m_db->delete_document(item.id());
     }
@@ -140,6 +158,8 @@ void AkonotesIndexer::remove(const Akonadi::Item &item)
 
 void AkonotesIndexer::remove(const Akonadi::Collection& collection)
 {
+    if (!m_db)
+        return;
     try {
         Xapian::Query query('C'+ QString::number(collection.id()).toStdString());
         Xapian::Enquire enquire(*m_db);
@@ -161,6 +181,8 @@ void AkonotesIndexer::move(const Akonadi::Item::Id& itemId,
                         const Akonadi::Entity::Id& from,
                         const Akonadi::Entity::Id& to)
 {
+    if (!m_db)
+        return;
     Xapian::Document doc;
     try {
         doc = m_db->get_document(itemId);
